# api/auth/swagger.yaml
basePath: /
definitions:
...
  authsdk.ClientInfo:
    properties:
      created_at:
        description: CreatedAt is the timestamp when the client was created (RFC3339
          format)
        type: string
      has_secret:
        description: HasSecret indicates whether this client has a secret (confidential
          client)
        type: boolean
      id:
        description: ID is the unique identifier for the client
        type: string
      name:
        description: Name is the human-readable name of the client
        type: string
      protected:
        description: Protected indicates whether this client is protected from deletion
        type: boolean
      scopes:
        description: Scopes is the list of scopes this client can grant
        items:
          type: string
        type: array
    type: object
  authsdk.CreateClientRequest:
    properties:
      confidential:
        description: |-
          Confidential indicates whether to create a confidential client with a secret.
          If true, a secret will be auto-generated and returned once.
          If false, creates a public client (no secret, cannot use client_credentials grant).
        type: boolean
      name:
        description: Name is the human-readable name for the client
        type: string
      scopes:
        description: Scopes is the list of scopes this client is authorized to grant
        items:
          type: string
        type: array
    type: object
  authsdk.CreateClientResponse:
    properties:
      client_id:
        description: ClientID is the unique identifier for the created client
        type: string
      client_secret:
        description: |-
          ClientSecret is the plaintext secret (only returned once at creation).
          Will be empty if no secret was provided in the request.
        type: string
    type: object
  authsdk.ErrorResponse:
    properties:
      error:
        description: Error is the OAuth2 error code (e.g., "invalid_request", "invalid_grant")
        type: string
      error_description:
        description: ErrorDescription is a human-readable description of the error
        type: string
    type: object
...
  authsdk.InviteRequest:
    properties:
      client_id:
        type: string
      expires_at:
        description: epoch time in seconds (1 day from creation if omitted)
        type: integer
      reusable:
        description: default false
        type: boolean
      role_id:
        description: Role to assign to the invited user
        type: string
    type: object
  authsdk.InviteResponse:
    properties:
      client_id:
        type: string
      expires_at:
        description: epoch time in seconds
        type: integer
      invite_token:
        type: string
    type: object
...
  authsdk.ListClientsResponse:
    properties:
      clients:
        items:
          $ref: '#/definitions/authsdk.ClientInfo'
        type: array
    type: object
...
  authsdk.RedeemInviteResponse:
    properties:
      user_id:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: AussieBroadWAN Team
    url: https://github.com/aussiebroadwan/bartab
  description: |-
    Attempting to be OAuth2-compliant for our authentication service providing token management with JWT-based access tokens.

    All tokens are signed using RS256 (RSA-SHA256) and can be verified using the JWKS endpoint.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: BarTab Authentication Service API
  version: 0.1.0
paths:
...
  /v1/clients:
    get:
      description: Returns all OAuth2 clients. Protected clients are flagged.
      parameters:
      - description: Bearer token with admin:read scope
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of clients
          schema:
            $ref: '#/definitions/authsdk.ListClientsResponse'
        "401":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "403":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List OAuth2 Clients
      tags:
      - Clients
    post:
      consumes:
      - application/json
      description: Creates a new OAuth2 client. If confidential=true, auto-generates
        a secret for client_credentials grant.
      parameters:
      - description: Bearer token with admin:write scope
        in: header
        name: Authorization
        required: true
        type: string
      - description: Client creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authsdk.CreateClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: client_id and client_secret (if confidential)
          schema:
            $ref: '#/definitions/authsdk.CreateClientResponse'
        "400":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "401":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "403":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create OAuth2 Client
      tags:
      - Clients
  /v1/clients/{id}:
    delete:
      description: Deletes an OAuth2 client by ID. Protected clients cannot be deleted.
      parameters:
      - description: Bearer token with admin:write scope
        in: header
        name: Authorization
        required: true
        type: string
      - description: Client ID (ULID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Client deleted successfully
        "400":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "401":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "403":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "404":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete OAuth2 Client
      tags:
      - Clients
  /v1/invites/mint:
    post:
      consumes:
      - application/json
      description: Mint a user invitation token for inviting new users for a specific
        client application. This is an admin-only operation.
      parameters:
      - description: Invite request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/authsdk.InviteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: invite_token, client_id, expires_at
          schema:
            $ref: '#/definitions/authsdk.InviteResponse'
        "400":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "401":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User Invitation Endpoint
      tags:
      - Invitations
  /v1/invites/redeem:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Redeem an invitation token to create a new user account
      parameters:
      - description: Invite token from mint endpoint
        in: formData
        name: invite_token
        required: true
        type: string
      - description: Desired username
        in: formData
        name: username
        required: true
        type: string
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      - description: Client ID the invite was issued for
        in: formData
        name: client_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: user_id, username
          schema:
            $ref: '#/definitions/authsdk.RedeemInviteResponse'
        "400":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
        "500":
          description: error, error_description
          schema:
            $ref: '#/definitions/authsdk.ErrorResponse'
      summary: Redeem Invitation Endpoint
      tags:
      - Invitations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Format: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"